requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn==1.5.2
joblib
python-dotenv
gunicorn